+++ source/geh_common/src/geh_common/infrastructure/write_csv.py
@@ -1,9 +1,8 @@
 import random
 import string
-import zipfile
 from dataclasses import dataclass
 from pathlib import Path
-from typing import Any, Callable
+from typing import Callable
 
 from pyspark.sql import DataFrame, Window
 from pyspark.sql import functions as F
@@ -44,43 +43,6 @@ class FileInfo:
     source: Path
     destination: Path
     temporary: Path
-
-
-def create_zip_file(path: str | Path, dbutils: Any, tmpdir: str | Path = Path("tmp")) -> Path:
-    """Create a zip file from a list of files and saves it to the specified path.
-
-    Notice that we have to create the zip file in /tmp and then move it to the desired
-    location. This is done as `direct-append` or `non-sequential` writes are not
-    supported in Databricks.
-
-    Args:
-        path (str | Path): The path to the directory containing the files to zip.
-        dbutils: The Databricks utilities object.
-        tmpdir (str | Path, optional): The temporary directory to write the zip file to. Defaults to "tmp".
-
-    Raises:
-        Exception: If there are no files to zip.
-
-    Returns:
-        Path: The path to the created zip file.
-    """
-    output_path = Path(path)
-    if not output_path.exists():
-        raise ValueError(f"'{output_path}' does not exist")
-    if not output_path.is_dir():
-        raise ValueError(f"'{output_path}' is not a directory")
-    zip_output_path = output_path.with_suffix(".zip")
-    files_to_zip = [Path(f) for f in dbutils.fs.ls(str(output_path))]
-    if len(files_to_zip) == 0:
-        raise FileNotFoundError(f"No files found in {output_path}")
-    tmp_path = Path(tmpdir) / zip_output_path.name
-    Path(tmp_path).parent.mkdir(parents=True, exist_ok=True)
-    with zipfile.ZipFile(tmp_path, "w", zipfile.ZIP_DEFLATED) as ref:
-        for fp in files_to_zip:
-            file_name = fp.name
-            ref.write(fp, arcname=file_name)
-    dbutils.fs.mv(f"file:{tmp_path}", zip_output_path)
-    return zip_output_path
 
 
 def write_csv_files(
@@ -133,11 +95,34 @@ def write_csv_files(
     return content
 
 
+def get_partition_information(path) -> dict[str, str]:
+    """Extract partition information from a file path.
+
+    Args:
+        path (str or Path): The file path from which to extract partition information.
+            The path should contain partition information in the format "key=value".
+            For example, "/tmp/part=1/part2=2/test" would yield {"part": "1", "part2": "2"}.
+
+    Raises:
+        ValueError: If the path contains an invalid partition format (e.g., "key=value=value").
+            This error occurs when there are too many values to unpack from the partition string.
+
+    Returns:
+        dict[str, str]: A dictionary with partition names as keys and their values.
+    """
+    partitions = {}
+    for p in Path(path).parts:
+        if "=" in p:
+            key, value = p.split("=", 2)
+            partitions[key] = value
+    return partitions
+
+
 def _get_file_info(
     result_output_path: str | Path,
     spark_output_path: str | Path,
     tmpdir: str | Path,
-    file_name_factory: FileNameCallbackType = _default_file_name_callback,
+    file_name_factory: FileNameCallbackType,
 ) -> list[FileInfo]:
     """Get file information for the files to be zipped.
 
@@ -258,26 +243,3 @@ def _merge_content(file_info: list[FileInfo], headers: list[str]) -> list[Path]:
                     fh_destination.write(fh_temporary.read())
 
     return list(destinations.keys())
-
-
-def get_partition_information(path) -> dict[str, str]:
-    """Extract partition information from a file path.
-
-    Args:
-        path (str or Path): The file path from which to extract partition information.
-            The path should contain partition information in the format "key=value".
-            For example, "/tmp/part=1/part2=2/test" would yield {"part": "1", "part2": "2"}.
-
-    Raises:
-        ValueError: If the path contains an invalid partition format (e.g., "key=value=value").
-            This error occurs when there are too many values to unpack from the partition string.
-
-    Returns:
-        dict[str, str]: A dictionary with partition names as keys and their values.
-    """
-    partitions = {}
-    for p in Path(path).parts:
-        if "=" in p:
-            key, value = p.split("=", 2)
-            partitions[key] = value
-    return partitions
